requirements.txt
streamlit>=1.30.0
numpy
soundfile
pydub
plotly
scipy
pandas
streamlit-audiorecorder
noisereduce
